w3c-reports/12092014/efl/runned-tests/html/webappapis/scripting/processing-model-2/compile-error-in-attribute-expected.txt

CONSOLE MESSAGE: line 25: SyntaxError: Unexpected token ')'

PASS window.onerror - compile error in attribute 
PASS window.onerror - compile error in attribute (column) 
